requirements.txt
avro-python3==1.8.2
confluent-kafka==1.3.0